+++ wcmatch/wcmatch.py
@@ -28,8 +28,8 @@ from . import util
 
 __all__ = (
     "FORCECASE", "IGNORECASE", "RAWCHARS", "FILEPATHNAME", "DIRPATHNAME",
-    "EXTMATCH", "GLOBSTAR", "BRACE", "MINUSNEGATE",
-    "F", "I", "R", "P", "E", "G", "M", "DP", "FP",
+    "EXTMATCH", "GLOBSTAR", "BRACE", "MINUSNEGATE", "SYMLINKS", "HIDDEN", "RECURSIVE"
+    "F", "I", "R", "P", "E", "G", "M", "DP", "FP", "SL", "HD", "RV",
     "WcMatch"
 )